requirements.txt
python-dotenv
flask
flask-cors
flask-restful
boto3
bcrypt
